// app/frontend/src/context/api/models/TodoDetailDto.ts
export interface TodoDetailDto {
    /**
     * 
     * @type {string}
     * @memberof TodoDetailDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TodoDetailDto
     */
    summary: string;
    /**
     * 
     * @type {string}
     * @memberof TodoDetailDto
     */
    description?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof TodoDetailDto
     */
    dueAt?: Date | null;
    /**
     * 
     * @type {TodoPriority}
     * @memberof TodoDetailDto
     */
    priority: TodoPriority;
    /**
     * 
     * @type {TodoState}
     * @memberof TodoDetailDto
     */
    state: TodoState;
    /**
     * 
     * @type {ProjectDetailDto}
     * @memberof TodoDetailDto
     */
    project: ProjectDetailDto;
    /**
     * 
     * @type {SectionDetailDto}
     * @memberof TodoDetailDto
     */
    section?: SectionDetailDto;
}